refactor: extract admin dashboard stats endpoint into dedicated router
